==> notebooks/03_preprocessing.ipynb <==
## <center style="color:red">**News Classifier**</center>

### <center>**ML Pipeline with Sentence Transformers, ChromaDB & Airflow**</center>

Ce projet vise à construire un pipeline complet de classification de textes permettant de catégoriser automatiquement des articles d’actualité dans 4 classes :

- World
- Sports
- Business
- Sci/Tech

Le pipeline utilise des techniques NLP avancées, des embeddings vectoriels, une base de données vectorielle et une orchestration via Apache Airflow.

<br>

### <span style="color:green">**Prétraîtement des Données :**</span>

#### <span style="color:orange">**1. Charger les Données :**</span>

import pandas as pd

df_train = pd.read_csv("../data/clean/train.csv")

df_test = pd.read_csv("../data/clean/test.csv")

print("Données Chargées avec Succès !")

df_train.head()

df_test.head()

#### <span style="color:orange">**2. Normaliser le Dataset :**</span>

df_train["text"] = df_train["text"].apply(lambda row : str(row).lower())

df_train.head()

df_test["text"] = df_test["text"].apply(lambda row : str(row).lower())

df_test.head()

#### <span style="color:orange">**3. Tokenisation :**</span>

import nltk
from nltk.tokenize import word_tokenize

nltk.download("punkt")
nltk.download("punkt_tab")

df_train["word_tokens"] = df_train["text"].apply(lambda x: word_tokenize(x) if isinstance(x, str) else [])

df_train.head()

df_test["word_tokens"] = df_test["text"].apply(lambda x: word_tokenize(x) if isinstance(x, str) else [])

df_test.head()

#### <span style="color:orange">**4. Retirer les Stopwords :**</span>

import nltk
from nltk.corpus import stopwords

nltk.download("stopwords")

stop_words = set(stopwords.words('english'))

df_train["tokens"] = df_train["word_tokens"].apply(
    lambda tokens : [token for token in tokens if token not in stop_words]
)

df_train.head()

df_test["tokens"] = df_test["word_tokens"].apply(
    lambda tokens : [token for token in tokens if token not in stop_words]
)

df_test.head()

#### <span style="color:orange">**5. Supprimer la Ponctuation & Caractères Spéciaux :**</span>

import re

df_train["clean_tokens"] = df_train["tokens"].apply(
    lambda tokens: [token for token in tokens if re.match(r"^[a-z]+$", token)]
)

df_train.head()

df_test["clean_tokens"] = df_test["tokens"].apply(
    lambda tokens: [token for token in tokens if re.match(r"^[a-z]+$", token)]
)

df_test.head()

#### <span style="color:orange">**6.  :**</span>

df_train["clean_text"] = df_train["clean_tokens"].apply(
    lambda token : " ".join(token)
)

df_train.head()

df_test["clean_text"] = df_test["clean_tokens"].apply(
    lambda token : " ".join(token)
)

df_test.head()

#### <span style="color:orange">**7. Sauvegarder les Datasets :**</span>

df_train = df_train[["text","clean_text","label"]]

df_train.head()

df_train.to_csv("../data/processed/train.csv", index=False)

print("Dataset Enregistré avec Succès !")

df_test = df_test[["text","clean_text","label"]]

df_test.head()

df_test.to_csv("../data/processed/test.csv", index=False)

print("Dataset Enregistré avec Succès !")
