==> notebooks/04_embedding.ipynb <==
## <center style="color:red">**News Classifier**</center>

### <center>**ML Pipeline with Sentence Transformers, ChromaDB & Airflow**</center>

Ce projet vise à construire un pipeline complet de classification de textes permettant de catégoriser automatiquement des articles d’actualité dans 4 classes :

- World
- Sports
- Business
- Sci/Tech

Le pipeline utilise des techniques NLP avancées, des embeddings vectoriels, une base de données vectorielle et une orchestration via Apache Airflow.

<br>

### <span style="color:green">**Embeddings & Vectorisation :**</span>

#### <span style="color:orange">**1. Charger les Données :**</span>

import pandas as pd

df_train = pd.read_csv("../data/processed/train.csv")

print("Données Chargées avec Succès !")

df_train.head()

df_test = pd.read_csv("../data/processed/test.csv")

print("Données Chargées avec Succès !")

df_test.head()

#### <span style="color:orange">**2. Encoder les Textes d’Entraînement et Test :**</span>

##### **2.1. Sentence Transformer :**

Un `Sentence Transformer` est un modèle de langage basé sur l’architecture **Transformer**, spécialement conçu pour produire des représentations vectorielles (**embeddings**) de phrases plutôt que de simples mots. 

Contrairement aux modèles classiques qui analysent chaque mot séparément, un `Sentence Transformer` capture le sens global d’une phrase, sa structure et ses relations sémantiques. Cela permet de :

- Comparer facilement deux phrases (similarité).

- Regrouper des textes proches.

- Faire de la recherche intelligente, du clustering, ou encore d’alimenter des modèles de classification avec des représentations beaucoup plus riches et précises. 

En pratique, il est largement utilisé dans des tâches de NLP comme la détection de doublons, le question-réponse, le résumé, la classification de textes avec peu d’exemples (few-shot), et bien plus.

Pour installer Sentence Transformers, voici la commande officielle :

```bash
pip install sentence-transformers
```

import sentence_transformers

print("Version : ", sentence_transformers.__version__)

##### **2.2. Charger le Modèle `paraphrase-multilingual-MiniLM-L12-v2` :**

Le `paraphrase-multilingual-MiniLM-L12-v2` est un modèle de la famille **Sentence Transformers**, entraîné pour générer des embeddings de phrases capables de mesurer la similarité sémantique entre textes. 

Ce modèle est multilingue, couvrant plus de 50 langues, ce qui lui permet de comprendre et comparer des phrases écrites dans différentes langues tout en gardant un espace vectoriel commun. 

Il est basé sur **MiniLM**, une architecture Transformer légère et rapide, composée ici de 12 couches, offrant un excellent compromis entre performance, vitesse et taille du modèle. 

En pratique, il est utilisé pour des tâches comme la détection de paraphrases, la recherche sémantique, le clustering de textes ou la classification, surtout lorsque l’on travaille sur des données multilingues.

from sentence_transformers import SentenceTransformer

model = SentenceTransformer("paraphrase-multilingual-MiniLM-L12-v2")

##### **2.3. Encoder les Textes :**

texts_train = df_train["text"].tolist()
texts_test = df_test["text"].tolist()

embeddings_train = model.encode(texts_train, normalize_embeddings=True, show_progress_bar=True)

embeddings_train

embeddings_test = model.encode(texts_test, normalize_embeddings=True, show_progress_bar=True)

embeddings_test

#### <span style="color:orange">**3. Sauvegarder les Datasets après Embedding :**</span>

df_train["text_embedding"] = embeddings_train.tolist()

df_train["id"] = "train_art_" + df_train.index.astype(str)

df_train.to_pickle("../metadatas/train.pkl")

print("Train Dataset Sauvegardé avec Succès !")

df_test["text_embedding"] = embeddings_test.tolist()

df_test["id"] = "test_art_" + df_test.index.astype(str)

df_test.to_pickle("../metadatas/test.pkl")

print("Test Dataset Sauvegardé avec Succès !")
