==> notebooks/05_chromaDB.ipynb <==
## <p style="color:red;text-align:center">**News Classifier**</p>

### <p style="text-align:center">**ML Pipeline with Sentence Transformers, ChromaDB & Airflow**</p>

Ce projet vise à construire un pipeline complet de classification de textes permettant de catégoriser automatiquement des articles d’actualité dans 4 classes :

- World
- Sports
- Business
- Sci/Tech

Le pipeline utilise des techniques NLP avancées, des embeddings vectoriels, une base de données vectorielle et une orchestration via Apache Airflow.

<br>

### <span style="color:green">**Stockage dans ChromaDB :**</span>

#### <span style="color:orange">**1. Charger les Données :**</span>

import pandas as pd

df_train = pd.read_pickle("../metadatas/train.pkl")

print("Train Dataset Chargée avec Succès !")

df_train.head()

import pandas as pd

df_test = pd.read_pickle("../metadatas/test.pkl")

print("Test Dataset Chargée avec Succès !")

df_test.head()

#### <span style="color:orange">**2. Installer & configurer `ChromaDB` :**</span>

`ChromaDB` est une **base de données vectorielle** open-source et performante, conçue spécifiquement pour servir de "mémoire à long terme" aux applications d'intelligence artificielle modernes. 

Contrairement aux bases de données classiques qui stockent du texte brut et cherchent des mots-clés exacts, `ChromaDB` stocke des **embeddings** (des vecteurs numériques représentant le sens sémantique d'un texte, d'une image ou d'un son) ainsi que leurs métadonnées associées. 

Son utilité principale réside dans sa capacité à effectuer des **recherches de similarité** ultra-rapides : elle permet à un modèle d'IA de retrouver instantanément les informations les plus pertinentes contextuellement par rapport à une requête donnée, ce qui la rend indispensable pour alimenter des Chatbots (RAG), des systèmes de recommandation ou des moteurs de recherche sémantique.

Pour installer ChromaDB, on utilise simplement :

```bash
pip install chromadb
```

import chromadb

print("Version :", chromadb.__version__)

#### <span style="color:orange">**3. Créer les Collections :**</span>

client = chromadb.PersistentClient(path="../chroma_db")

collection_train = client.create_collection(name="train_news", metadata={"hnsw:space" : "cosine"})
collection_test = client.create_collection(name="test_news", metadata={"hnsw:space" : "cosine"})

#### <span style="color:orange">**4. Insérer Embeddings, Identifiants et Labels :**</span>

batch_size = 5000

for i in range(0, len(df_train), batch_size):
    batch = df_train.iloc[i:i + batch_size]

    batch_ids = batch["id"].astype(str).tolist()
    batch_embeddings = batch["text_embedding"].tolist()
    batch_metadata = batch[["label", "text"]].to_dict("records")

    collection_train.add(
        ids=batch_ids,
        embeddings=batch_embeddings,
        metadatas=batch_metadata
    )

print(f"Nombre des Items dans la Collection d'Entraînement : {collection_train.count()}")

batch_size = 5000

for i in range(0, len(df_test), batch_size):
    batch = df_test.iloc[i:i + batch_size]

    batch_ids = batch["id"].astype(str).tolist()
    batch_embeddings = batch["text_embedding"].tolist()
    batch_metadata = batch[["label", "text"]].to_dict("records")

    collection_test.add(
        ids=batch_ids,
        embeddings=batch_embeddings,
        metadatas=batch_metadata
    )

print(f"Nombre des Items dans la Collection de Test : {collection_test.count()}")

train_item = collection_train.peek(limit=1)

print("\nTrain Item : ")
print(f"- ID: {train_item['ids'][0]}")
print(f"- Label: {train_item['metadatas'][0]['label']}")
print(f"- Text: {train_item['metadatas'][0]['text'][:50]}...")
print(f"- Embedding : {train_item['embeddings'][0][:10]}...")
print(f"- Embedding length: {len(train_item['embeddings'][0])}")

test_item = collection_test.peek(limit=1)

print("\Test Item : ")
print(f"- ID: {test_item['ids'][0]}")
print(f"- Label: {test_item['metadatas'][0]['label']}")
print(f"- Text: {test_item['metadatas'][0]['text'][:50]}...")
print(f"- Embedding : {test_item['embeddings'][0][:10]}...")
print(f"- Embedding length: {len(test_item['embeddings'][0])}")
